Setted up celery
