+++ backend/.env
@@ -4,7 +4,7 @@ JWT_ALGORITHM = "HS256"
 ACCES_JWT_EXPIRY_SECONDS = "3600"
 REFRESH_JWT_EXPIRY_SECONDS = "36000"
 
-POST_RATING_EXPIRATION = "3600"
+POST_RATING_EXPIRATION = "5"
 
 QUERY_PARAM_MAX_L = "500"
 
@@ -24,6 +24,10 @@ HISTORY_POSTS_TO_TAKE_INTO_RELATED = "30"
 
 PASSWORD_MIN_L = "8"
 PASSWORD_MAX_L = "32"
+
+# CELERY | Call in backend directory
+# celery -A post_popularity_rate_celery.popularity_rate  worker --loglevel=info -P gevent
+# celery -A post_popularity_rate_celery.celery_main beat -l info
 
 #REDIS
     REDIS_PORT = "6379"
