+++ backend/databases_manager/postgres_manager/postgres_manager.py
@@ -12,7 +12,7 @@ from uuid import UUID
 
 Models = TypeVar("Models", bound=Base)
 
-FEED_MAX_POSTS_LOADED = int(getenv("FEED_MAX_POSTS_LOADED"))
+FEED_MAX_POSTS_LOADED = int(getenv("FEED_MAX_POSTS_LOAD"))
 N_MAX_FRESH_POSTS_TO_MIX = int(getenv("N_MAX_FRESH_POSTS_TO_MIX"))
 
 MAX_FOLLOWED_POSTS_TO_SHOW = int(getenv("MAX_FOLLOWED_POSTS_TO_SHOW"))
